requirements.txt
config-with-yaml==0.1.0
forcediphttpsadapter==1.0.1
netaddr==0.8.0
progressbar2==3.51.4
pydash==4.8.0
pyinstaller==4.0
requests==2.24.0
